fixed signed vs. unsigned warning
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9787 d7cf8633-e32d-0410-b094-e92efae38249

+++ console.c
@@ -1384,7 +1384,7 @@ void Con_DrawInput (void)
 	y = (int)strlen(text);
 
 	// append enoug nul-bytes to cover the utf8-versions of the cursor too
-	for (i = y; i < y + 4 && i < sizeof(editlinecopy); ++i)
+	for (i = y; i < y + 4 && i < (int)sizeof(editlinecopy); ++i)
 		text[i] = 0;
 
 	// add the cursor frame
